codi bonic per al tesseraact i el easyOCR

+++ matricules/charSegment_pipeline.py
@@ -1,6 +1,6 @@
 import cv2
 
-def segmentChars(roi):
+def segmentChars(roi, debug=False):
 
     # Passem la imatge a gris
     roi_gray = cv2.cvtColor(roi, cv2.COLOR_BGR2GRAY)
@@ -22,10 +22,10 @@ def segmentChars(roi):
         if h > 15 and w > 15 and (w < roi.shape[1]/3):
             #cv2.rectangle(roi, (x-2, y-5), (x-2 + w+5, y-5 + h+10), (0, 255, 0), 2)
             letters_bboxes.append((x-2,y-5,w+2,h+10))
-    
-    cv2.imshow('Image with Bounding Boxes', roi)
-    cv2.waitKey(0)
-    cv2.destroyAllWindows()
+    if debug:
+        cv2.imshow('Image with Bounding Boxes', roi)
+        cv2.waitKey(0)
+        cv2.destroyAllWindows()
 
     return letters_bboxes
 """
